requirements.txt
configobj
flask
flask-bootstrap
python-redmine
